fix(runtime): tolerate platform-narrower stat field widths in lease-mutex
st_dev/st_nlink are narrower than u64 on some platforms (e.g. Linux x86_64)
but not others (e.g. macOS), so the identical 'as u64' cast is flagged as
unnecessary by clippy on Linux while remaining a real narrowing conversion
on Darwin. Scope an allow to this cross-platform-portable stat conversion.

+++ crates/omx-runtime/src/lease_mutex.rs
@@ -367,10 +367,11 @@ fn stat_fd(fd: i32, label: &str) -> Result<Identity, String> {
 }
 
 #[cfg(unix)]
+#[allow(clippy::unnecessary_cast)]
 fn identity_from_stat(st: &libc::stat) -> Identity {
     Identity {
         dev: st.st_dev as u64,
-        ino: st.st_ino,
+        ino: st.st_ino as u64,
         nlink: st.st_nlink as u64,
         is_dir: (st.st_mode & libc::S_IFMT) == libc::S_IFDIR,
         is_regular: (st.st_mode & libc::S_IFMT) == libc::S_IFREG,
